+++ blog-dao/src/main/resources/com/zuoxiaolong/blog/mapper/WebUserMapper.xml
@@ -21,11 +21,11 @@
         where id = #{id,jdbcType=INTEGER}
     </select>
 
-    <select id="selectByWebUserId" resultMap="BaseResultMap" parameterType="java.lang.Integer">
+    <select id="selectByUsername" resultMap="BaseResultMap" parameterType="java.lang.String">
         select
         <include refid="Base_Column_List"/>
         from web_user
-        where web_user_id = #{webUserId,jdbcType=INTEGER}
+        where username = #{username,jdbcType=INTEGER}
     </select>
 
     <delete id="deleteByPrimaryKey" parameterType="java.lang.Integer">
